Move HomePage to views folder
